+++ .gitignore
@@ -49,3 +49,6 @@ build/
 
 # vscode
 .vscode/
+
+# settings
+.settings.json
